MIUW/requirements.txt
django
djangorestframework
requests
spotipy
django-cors-headers